backend/requirements.txt
fastapi
uvicorn
python-multipart
torch --index-url https://download.pytorch.org/whl/cpu
transformers
pytesseract
Pillow
numpy